+++ sdk/storage/azure_storage_blob/tests/page_blob_client.rs
@@ -3,13 +3,10 @@
 
 use azure_core::http::{headers::CONTENT_TYPE, RequestContent, StatusCode};
 use azure_core_test::{recorded, TestContext};
-use azure_storage_blob::{
-    format_page_range,
-    models::{
-        BlobClientGetPropertiesResultHeaders, BlobType, PageBlobClientCreateOptions,
-        PageBlobClientSetSequenceNumberOptions, PageBlobClientSetSequenceNumberResultHeaders,
-        SequenceNumberActionType,
-    },
+use azure_storage_blob::models::{
+    BlobClientGetPropertiesResultHeaders, BlobType, HttpRange, PageBlobClientCreateOptions,
+    PageBlobClientSetSequenceNumberOptions, PageBlobClientSetSequenceNumberResultHeaders,
+    SequenceNumberActionType,
 };
 use azure_storage_blob_test::{get_blob_name, get_container_client, StorageAccount};
 use std::error::Error;
@@ -69,7 +66,7 @@ async fn test_upload_page(ctx: TestContext) -> Result<(), Box<dyn Error>> {
         .upload_pages(
             RequestContent::from(data.clone()),
             512,
-            format_page_range(0, 512)?,
+            HttpRange::new(0, 512).into(),
             None,
         )
         .await?;
@@ -98,13 +95,13 @@ async fn test_clear_page(ctx: TestContext) -> Result<(), Box<dyn Error>> {
         .upload_pages(
             RequestContent::from(data),
             512,
-            format_page_range(0, 512)?,
+            HttpRange::new(0, 512).into(),
             None,
         )
         .await?;
 
     page_blob_client
-        .clear_pages(format_page_range(0, 512)?, None)
+        .clear_pages(HttpRange::new(0, 512).into(), None)
         .await?;
 
     // Assert
@@ -133,7 +130,7 @@ async fn test_resize_blob(ctx: TestContext) -> Result<(), Box<dyn Error>> {
         .upload_pages(
             RequestContent::from(data.clone()),
             1024,
-            format_page_range(0, 1024)?,
+            HttpRange::new(0, 1024).into(),
             None,
         )
         .await;
@@ -146,7 +143,7 @@ async fn test_resize_blob(ctx: TestContext) -> Result<(), Box<dyn Error>> {
         .upload_pages(
             RequestContent::from(data.clone()),
             1024,
-            format_page_range(0, 1024)?,
+            HttpRange::new(0, 1024).into(),
             None,
         )
         .await?;
@@ -228,7 +225,7 @@ async fn test_upload_page_from_url(ctx: TestContext) -> Result<(), Box<dyn Error
         .upload_pages(
             RequestContent::from(data_b.clone()),
             512,
-            format_page_range(0, 512)?,
+            HttpRange::new(0, 512).into(),
             None,
         )
         .await?;
@@ -239,16 +236,16 @@ async fn test_upload_page_from_url(ctx: TestContext) -> Result<(), Box<dyn Error
         .upload_pages(
             RequestContent::from(data_a.clone()),
             512,
-            format_page_range(0, 512)?,
+            HttpRange::new(0, 512).into(),
             None,
         )
         .await?;
     page_blob_client_2
         .upload_pages_from_url(
             blob_client_1.url().as_str().into(),
-            format_page_range(0, data_b.len() as u64)?,
+            HttpRange::new(0, data_b.len() as u64).into(),
             data_b.len() as u64,
-            format_page_range(512, data_b.len() as u64)?,
+            HttpRange::new(512, data_b.len() as u64).into(),
             None,
         )
         .await?;
@@ -287,7 +284,7 @@ async fn test_get_page_ranges(ctx: TestContext) -> Result<(), Box<dyn Error>> {
         .upload_pages(
             RequestContent::from(data.clone()),
             512,
-            format_page_range(0, 512)?,
+            HttpRange::new(0, 512).into(),
             None,
         )
         .await?;